requirements.txt
django==2.0.2
djangorestframework==3.7.7
django-filter==1.1.0
djangorestframework-jwt==1.11.0
names==0.3.0
requests
mock
